pyproject.toml
[project]
name = "airtable"
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flake8>=7.2.0",
    "freezegun>=1.5.1",
    "keboola-component>=1.6.10",
    "keboola-http-client>=1.0.1",
    "keboola-utils>=1.1.0",
    "mock>=5.2.0",
    "pydantic>=2.11.3",
    "ruff>=0.11.5",
    "pandas>=2.0.0",
    "pyairtable>=2.1.0",
]

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | env
  | build
  | dist
  | data
)/
'''